refactor: consolidate user type options and default search values into shared configs
