+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/configs/BulkWitnessDepositionSignConfig.js
@@ -1,6 +1,4 @@
-const defaultSearchValues = {
-  caseTitle: "",
-};
+import { bulkInboxCaseTitleSearchDefaults as defaultSearchValues } from "./shared/bulkInboxCaseTitleDefaults";
 
 const limit = parseInt(sessionStorage.getItem("bulkWitnessDepositionSignlimit") || 10);
 const offset = parseInt(sessionStorage.getItem("bulkWitnessDepositionSignoffset") || 0);
