+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/configs/LitigantHomeConfig.js
@@ -1,9 +1,5 @@
-const defaultSearchValues = {
-  caseSearchText: "",
-  caseType: "NIA S138",
-  substage: "",
-  secondaryStage: "",
-};
+import { defaultSearchValues, userTypeOptions } from "./shared/employeeHomeSearchDefaults";
+export { userTypeOptions };
 
 export const CaseWorkflowState = {
   CASE_REASSIGNED: "CASE_REASSIGNED",
@@ -12,112 +8,6 @@ export const CaseWorkflowState = {
   CASE_ADMITTED: "CASE_ADMITTED",
   PENDING_ADMISSION: "PENDING_ADMISSION",
 };
-
-export const userTypeOptions = [
-  {
-    code: "LITIGANT",
-    name: "LITIGANT_TEXT",
-    showBarDetails: false,
-    isVerified: false,
-    role: [
-      "CASE_CREATOR",
-      "CASE_EDITOR",
-      "CASE_VIEWER",
-      "EVIDENCE_CREATOR",
-      "EVIDENCE_VIEWER",
-      "EVIDENCE_EDITOR",
-      "APPLICATION_CREATOR",
-      "APPLICATION_VIEWER",
-      "HEARING_VIEWER",
-      "ORDER_VIEWER",
-      "SUBMISSION_CREATOR",
-      "SUBMISSION_RESPONDER",
-      "SUBMISSION_DELETE",
-      "TASK_VIEWER",
-      "ADVOCATE_VIEWER",
-      "PENDING_TASK_CREATOR",
-      "BAIL_BOND_CREATOR",
-      "BAIL_BOND_VIEWER",
-      "BAIL_BOND_EDITOR",
-    ],
-    subText: "LITIGANT_SUB_TEXT",
-  },
-  {
-    code: "ADVOCATE",
-    name: "ADVOCATE_TEXT",
-    showBarDetails: true,
-    isVerified: true,
-    hasBarRegistrationNo: true,
-    role: [
-      "ADVOCATE_ROLE",
-      "CASE_CREATOR",
-      "CASE_EDITOR",
-      "CASE_VIEWER",
-      "EVIDENCE_CREATOR",
-      "EVIDENCE_VIEWER",
-      "EVIDENCE_EDITOR",
-      "APPLICATION_CREATOR",
-      "APPLICATION_VIEWER",
-      "HEARING_VIEWER",
-      "ORDER_VIEWER",
-      "SUBMISSION_CREATOR",
-      "SUBMISSION_RESPONDER",
-      "SUBMISSION_DELETE",
-      "TASK_VIEWER",
-      "USER_REGISTER",
-      "ADVOCATE_VIEWER",
-      "ADVOCATE_APPLICATION_VIEWER",
-      "PENDING_TASK_CREATOR",
-      "BAIL_BOND_CREATOR",
-      "BAIL_BOND_VIEWER",
-      "BAIL_BOND_EDITOR",
-    ],
-    apiDetails: {
-      serviceName: "/advocate/v1/_create",
-      requestKey: "advocate",
-      AdditionalFields: ["barRegistrationNumber"],
-    },
-    subText: "ADVOCATE_SUB_TEXT",
-  },
-  {
-    code: "ADVOCATE_CLERK",
-    name: "ADVOCATE_CLERK_TEXT",
-    showBarDetails: true,
-    hasStateRegistrationNo: true,
-    isVerified: true,
-    role: [
-      "ADVOCATE_CLERK_ROLE",
-      "CASE_CREATOR",
-      "CASE_EDITOR",
-      "CASE_VIEWER",
-      "EVIDENCE_CREATOR",
-      "EVIDENCE_VIEWER",
-      "EVIDENCE_EDITOR",
-      "APPLICATION_CREATOR",
-      "APPLICATION_VIEWER",
-      "HEARING_VIEWER",
-      "ORDER_VIEWER",
-      "SUBMISSION_CREATOR",
-      "SUBMISSION_RESPONDER",
-      "SUBMISSION_DELETE",
-      "TASK_VIEWER",
-      "USER_REGISTER",
-      "ADVOCATE_VIEWER",
-      "ADVOCATE_APPLICATION_VIEWER",
-      "PENDING_TASK_CREATOR",
-      "BAIL_BOND_CREATOR",
-      "BAIL_BOND_VIEWER",
-      "BAIL_BOND_EDITOR",
-    ],
-    apiDetails: {
-      serviceName: "/advocate/clerk/v1/_create",
-      requestKey: "clerk",
-      AdditionalFields: ["stateRegnNumber"],
-    },
-
-    subText: "ADVOCATE_CLERK_SUB_TEXT",
-  },
-];
 
 export const TabLitigantSearchConfig = {
   tenantId: "pg",
