+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/configs/shared/bulkInboxCaseTitleDefaults.js
@@ -1,0 +1,4 @@
+/** Shared default search fields for bulk inbox flows that only filter by case title (Phase 3.4). */
+export const bulkInboxCaseTitleSearchDefaults = {
+  caseTitle: "",
+};
